Refactor and working parser
