--- src/main/java/antlr/treeFormatParser/generate/TreeFormat.interp
@@ -1,31 +1,0 @@
-token literal names:
-null
-':'
-null
-null
-null
-null
-null
-null
-
-token symbolic names:
-null
-COLON
-STRING
-MARKLESS_STRING
-NEWLINE
-WS
-INDENT
-DEDENT
-
-rule names:
-subtree
-child
-sibling
-wrapper
-node
-information
-
-
-atn:
-[3, 24715, 42794, 33075, 47597, 16764, 15335, 30598, 22884, 3, 9, 48, 4, 2, 9, 2, 4, 3, 9, 3, 4, 4, 9, 4, 4, 5, 9, 5, 4, 6, 9, 6, 4, 7, 9, 7, 3, 2, 3, 2, 7, 2, 17, 10, 2, 12, 2, 14, 2, 20, 11, 2, 3, 2, 3, 2, 3, 2, 7, 2, 25, 10, 2, 12, 2, 14, 2, 28, 11, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 4, 6, 4, 36, 10, 4, 13, 4, 14, 4, 37, 3, 5, 3, 5, 3, 6, 3, 6, 3, 6, 3, 6, 3, 7, 3, 7, 3, 7, 2, 2, 8, 2, 4, 6, 8, 10, 12, 2, 3, 3, 2, 4, 5, 2, 44, 2, 14, 3, 2, 2, 2, 4, 29, 3, 2, 2, 2, 6, 35, 3, 2, 2, 2, 8, 39, 3, 2, 2, 2, 10, 41, 3, 2, 2, 2, 12, 45, 3, 2, 2, 2, 14, 18, 5, 10, 6, 2, 15, 17, 7, 6, 2, 2, 16, 15, 3, 2, 2, 2, 17, 20, 3, 2, 2, 2, 18, 16, 3, 2, 2, 2, 18, 19, 3, 2, 2, 2, 19, 21, 3, 2, 2, 2, 20, 18, 3, 2, 2, 2, 21, 22, 5, 4, 3, 2, 22, 26, 5, 10, 6, 2, 23, 25, 7, 6, 2, 2, 24, 23, 3, 2, 2, 2, 25, 28, 3, 2, 2, 2, 26, 24, 3, 2, 2, 2, 26, 27, 3, 2, 2, 2, 27, 3, 3, 2, 2, 2, 28, 26, 3, 2, 2, 2, 29, 30, 7, 6, 2, 2, 30, 31, 7, 8, 2, 2, 31, 32, 5, 6, 4, 2, 32, 33, 7, 9, 2, 2, 33, 5, 3, 2, 2, 2, 34, 36, 5, 8, 5, 2, 35, 34, 3, 2, 2, 2, 36, 37, 3, 2, 2, 2, 37, 35, 3, 2, 2, 2, 37, 38, 3, 2, 2, 2, 38, 7, 3, 2, 2, 2, 39, 40, 5, 2, 2, 2, 40, 9, 3, 2, 2, 2, 41, 42, 5, 12, 7, 2, 42, 43, 7, 3, 2, 2, 43, 44, 5, 12, 7, 2, 44, 11, 3, 2, 2, 2, 45, 46, 9, 2, 2, 2, 46, 13, 3, 2, 2, 2, 5, 18, 26, 37]
