+++ src/main/java/antlr/treepatParser/TreepatVisitorImplementation.kt
@@ -1,7 +1,7 @@
 package antlr.treepatParser
 
-import antlr.treepatParser.generate.TreepatParser
-import antlr.treepatParser.generate.TreepatVisitor
+import antlr.treepatParser.generated.TreepatParser
+import antlr.treepatParser.generated.TreepatVisitor
 import ast.ASTNode
 import ast.Child
 import ast.Node
