+++ test.tef
@@ -1,0 +1,5 @@
+nodo1:et1
+	nodo2:et2
+	nodo3:et2
+		nodo4:et3
+	nodo5:et1
