+++ test.tp
@@ -1,11 +1,10 @@
 A
-   B
-        K
-   C
-        F
-        E*
-        (
-            @J
-            O
-        )#
-
+	B
+		K
+	C
+		F
+		E*
+		(
+			@J
+			O
+		)#
